requirements.txt
git+https://github.com/LIMBODS/FunPayAPI.git
python-telegram-bot==20.7
python-dotenv==1.0.0
aiosqlite==0.19.0
aiohttp==3.9.1
tenacity==8.2.3
